+++ tests/test_handler.py
@@ -1,8 +1,7 @@
-import json
-
 from logging import makeLogRecord
 
-from cee_syslog_handler import CeeSysLogHandler, JsonFormatter
+from cee_syslog_handler import CeeSysLogHandler
+
 
 def test_default_facility():
     record = makeLogRecord( {'name':'my.package.logger'} )
@@ -65,49 +64,3 @@ def test_custom_field_with_raising_str():
     handler = CeeSysLogHandler()
 
     assert '"_custom_field": "value could not be converted to str"' in handler.format(record)
-
-def test_json_formatter():
-    fmt = JsonFormatter()
-    record = makeLogRecord( {'name':'my.package.logger'} )
-    res = fmt.format(record)
-
-    assert '"facility": "my.package.logger' in res
-   
-    d = json.loads(res)
-    assert isinstance(d, dict)
-    for f in ["file", "line", "_function", "_pid", "_thread_name"]:
-        assert f in d
-
-def test_json_formatter_options():
-    fmt = JsonFormatter(debugging_fields=False)
-    record = makeLogRecord( {'name':'my.package.logger', 'special_field': 10} )
-    res = fmt.format(record)
-
-    d = json.loads(res)
-    assert isinstance(d, dict)
-    for f in ["file", "line", "_function", "_pid", "_thread_name"]:
-        assert f not in d
-    
-    assert d["_special_field"] == 10
-
-    fmt = JsonFormatter(datefmt='%Y-%m-%d',)
-    record = makeLogRecord( {'name':'my.package.logger', 'special_field': 10} )
-    res = fmt.format(record)
-
-    d = json.loads(res)
-    assert len(d["timestamp"]) == 10
-
-
-    fmt = JsonFormatter(extra_fields=False)
-    record = makeLogRecord( {'name':'my.package.logger', 'special_field': 10} )
-    res = fmt.format(record)
-
-    d = json.loads(res)
-    assert "_special_field" not in d
-
-
-
-
-
-
-
