+++ tests/test_named_logger.py
@@ -1,7 +1,6 @@
 # coding=utf-8
 
 import logging
-import unittest
 
 from cee_syslog_handler import NamedCeeLogger
 
@@ -10,22 +9,17 @@ _DUMMY_HOST = ('localhost', 1337)
 _DUMMY_PROTOCOL = 2
 
 
-def _make_log_record():
-    return logging.LogRecord('dummy_logger', logging.DEBUG, 'test.py', 13, 'Dummy message', (), None)
+def _check_format_adds_service_name(expected_service_name):
+    expected_json_field = '"_name": "{}"'.format(expected_service_name)
+    logger = NamedCeeLogger(_DUMMY_HOST, 2, expected_service_name)
+    record = logging.LogRecord('dummy_logger', logging.DEBUG, 'test.py', 13, 'Dummy message', (), None)
+
+    assert expected_json_field in logger.format(record)
 
 
-class NamedCeeLoggerTests(unittest.TestCase):
-    def _test_format_adds_service_name(self, expected_service_name):
-        expected_json_field = '"_name": "{}"'.format(expected_service_name)
-        logger = NamedCeeLogger(_DUMMY_HOST, 2, expected_service_name)
-        record = _make_log_record()
-        self.assertIn(expected_json_field, logger.format(record))
-
-    def test_format_adds_service_name(self):
-        """
-        Ensure the Cee entry created by format contains the name provided to the logger constructor
-        """
-        self._test_format_adds_service_name('My Fancy Service')
-        self._test_format_adds_service_name('Another Service')
-
-
+def test_format_adds_service_name():
+    """
+    Ensure the Cee entry created by format contains the name provided to the logger constructor
+    """
+    _check_format_adds_service_name('My Fancy Service')
+    _check_format_adds_service_name('Another Service')
